Split LinkWidget tests from plugin tests.
